+++ tests/test_parsers.py
@@ -20,7 +20,7 @@
 
 from pathlib import Path
 
-from chunked_scatter.chunked_scatter import BedRegion, file_to_regions
+from chunked_scatter.parsers import BedRegion, file_to_regions
 
 import pytest
 
@@ -44,8 +44,17 @@ def test_file_to_regions_dict():
     ]
 
 
+def test_file_to_regions_fai():
+    result = list(file_to_regions(datadir / "reference.fasta.fai"))
+    assert result == [
+        BedRegion("chr1", 0, 15624),
+        BedRegion("chrM", 0, 16569),
+        BedRegion("chr1_alt", 0, 6120)
+    ]
+
+
 def test_file_to_regions_wrong_ext(capsys):
     with pytest.raises(NotImplementedError,
-                       match="Only files with .bed or .dict extensions are "
-                             "supported."):
+                       match="Only files with .bed, .fai or .dict extensions "
+                             "are supported."):
         file_to_regions(Path("input"))