+++ CHANGELOG.rst
@@ -5,6 +5,12 @@ Changelog
 
 version 0.2.0-dev
 ---------------------------
++ The ``--input`` has now been abolished in favor of a positional input
+  argument.
++ The ``--prefix`` flag now has ``scatter-`` as default value.
++ Added a ``--print-paths`` or ``-P`` flag that prints the paths of the output
+  files to stdout.
++ Added support for fasta index (``.fai``) files.
 + Added a ``--split-contigs`` flag to allow contigs to be split up over
   multiple files.
 + Added new ``scatter-regions`` command that is a replacement for